+++ packages/nesd/lib/ui/settings/graphics/graphics_settings.dart
@@ -1,24 +1,34 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:nesd/ui/settings/graphics/border_switch.dart';
+import 'package:nesd/ui/settings/graphics/pixel_aspect_ratio_dropdown.dart';
+import 'package:nesd/ui/settings/graphics/pixel_aspect_ratio_slider.dart';
 import 'package:nesd/ui/settings/graphics/renderer_selector.dart';
 import 'package:nesd/ui/settings/graphics/scaling_dropdown.dart';
-import 'package:nesd/ui/settings/graphics/stretch_switch.dart';
+import 'package:nesd/ui/settings/settings.dart';
 import 'package:nesd/ui/settings/settings_tab.dart';
 
-class GraphicsSettings extends StatelessWidget {
+class GraphicsSettings extends ConsumerWidget {
   const GraphicsSettings({super.key});
 
   @override
-  Widget build(BuildContext context) {
-    return const SettingsTab(
+  Widget build(BuildContext context, WidgetRef ref) {
+    final pixelAspectRatio = ref.watch(
+      settingsControllerProvider.select((s) => s.pixelAspectRatio),
+    );
+
+    return SettingsTab(
       index: 1,
       child: SingleChildScrollView(
         child: Column(
           children: [
-            RendererSelector(),
-            StretchSwitch(),
-            BorderSwitch(),
-            ScalingDropdown(),
+            const RendererSelector(),
+            const BorderSwitch(),
+            const ScalingDropdown(),
+            const PixelAspectRatioDropdown(),
+            PixelAspectRatioSlider(
+              enabled: pixelAspectRatio == PixelAspectRatio.custom,
+            ),
           ],
         ),
       ),